+++ src/lib/constants/tracks.js
@@ -17,3 +17,40 @@ export const TRACK_TYPES = [
     STACK,
     TEXT,
 ];
+
+/**
+Visualization types supported by pileup.js
+Options defined in https://github.com/akmorrow13/pileup.js/blob/master/src/main/pileup.js
+*/
+export const PILEUP_VIZ_TYPES = [
+    'coverage',
+    'genome',
+    'genes',
+    'features',
+    'idiogram',
+    'location',
+    'scale',
+    'variants',
+    'genotypes',
+    'pileup',
+];
+
+/**
+Data source types supported by pileup.js
+Options defined in https://github.com/akmorrow13/pileup.js/blob/master/src/main/pileup.js
+*/
+export const PILEUP_SOURCE_TYPES = [
+    'bam',
+    'alignmentJson',
+    'variantJson',
+    'featureJson',
+    'idiogramJson',
+    'cytoBand',
+    'vcf',
+    'twoBit',
+    'bigBed',
+    'GAReadAlignment',
+    'GAVariant',
+    'GAFeature',
+    'GAGene',
+];